+++ src/src/Helpers/Scheduler_TaskDeviceTimer.cpp
@@ -1,0 +1,93 @@
+#include "../Helpers/Scheduler.h"
+
+
+#include "../DataStructs/Scheduler_TaskDeviceTimerID.h"
+#include "../DataStructs/TimingStats.h"
+#include "../ESPEasyCore/Controller.h"
+#include "../Globals/Settings.h"
+#include "../Helpers/DeepSleep.h"
+
+/*********************************************************************************************\
+* Task Device Timer
+* This is the interval set in a plugin to get a new reading.
+* These timers will re-schedule themselves as long as the plugin task is enabled.
+* When the plugin task is initialized, a call to schedule_task_device_timer_at_init
+* will bootstrap this sequence.
+\*********************************************************************************************/
+void ESPEasy_Scheduler::schedule_task_device_timer_at_init(unsigned long task_index) {
+  unsigned long runAt = millis();
+
+  if (!isDeepSleepEnabled()) {
+    // Deepsleep is not enabled, add some offset based on the task index
+    // to make sure not all are run at the same time.
+    // This scheduled time may be overriden by the plugin's own init.
+    runAt += (task_index * 37) + 100;
+  } else {
+    runAt += (task_index * 11) + 10;
+  }
+  schedule_task_device_timer(task_index, runAt);
+}
+
+// Typical use case is to run this when all needed connections are made.
+void ESPEasy_Scheduler::schedule_all_task_device_timers() {
+  for (taskIndex_t task = 0; task < TASKS_MAX; task++) {
+    schedule_task_device_timer_at_init(task);
+  }
+}
+
+void ESPEasy_Scheduler::schedule_task_device_timer(unsigned long task_index, unsigned long runAt) {
+  /*
+     String log = F("schedule_task_device_timer: task: ");
+     log += task_index;
+     log += F(" @ ");
+     log += runAt;
+     if (Settings.TaskDeviceEnabled[task_index]) {
+      log += F(" (enabled)");
+     }
+     addLog(LOG_LEVEL_INFO, log);
+   */
+
+  if (!validTaskIndex(task_index)) { return; }
+
+  const deviceIndex_t DeviceIndex = getDeviceIndex_from_TaskIndex(task_index);
+
+  if (!validDeviceIndex(DeviceIndex)) { return; }
+
+  // TD-er: Tasks without a timer or optional timer set to 0 should still be able to call PLUGIN_READ
+  // For example to schedule a read from the PLUGIN_TEN_PER_SECOND when a new value is ready.
+
+  /*
+     if (!Device[DeviceIndex].TimerOption) { return; }
+
+     if (Device[DeviceIndex].TimerOptional && (Settings.TaskDeviceTimer[task_index] == 0)) {
+      return;
+     }
+   */
+
+  if (Settings.TaskDeviceEnabled[task_index]) {
+    const TaskDeviceTimerID timerID(task_index);
+    setNewTimerAt(timerID, runAt);
+  }
+}
+
+void ESPEasy_Scheduler::reschedule_task_device_timer(unsigned long task_index, unsigned long lasttimer) {
+  if (!validTaskIndex(task_index)) { return; }
+  unsigned long newtimer = Settings.TaskDeviceTimer[task_index];
+
+  if (newtimer != 0) {
+    newtimer = lasttimer + (newtimer * 1000);
+    schedule_task_device_timer(task_index, newtimer);
+  }
+}
+
+void ESPEasy_Scheduler::process_task_device_timer(SchedulerTimerID timerID, unsigned long lasttimer) {
+  const TaskDeviceTimerID *tmp = reinterpret_cast<const TaskDeviceTimerID *>(&timerID);
+  const taskIndex_t task_index = tmp->getTaskIndex();
+
+  if (!validTaskIndex(task_index)) { return; }
+  START_TIMER;
+  struct EventStruct TempEvent(task_index);
+
+  SensorSendTask(&TempEvent, 0, lasttimer);
+  STOP_TIMER(SENSOR_SEND_TASK);
+}
